* use the MaterialShapeDrawable instead GradientDrawable for the background * move typeface to general AbstratDrawerItem * move some helper getter methods to the AbstractDrawerItem * allow to provide a custom ShapeAppearanceModel to the item implementation
